+++ packages/explorable-explanations/src/components/box.js
@@ -18,7 +18,7 @@
  */
 import app from '../protectedAudience/app'; // @todo Seperate for the protected audience.
 import config from '../protectedAudience/config';
-import utils from '../protectedAudience/lib/utils';
+import { addCanvasEventListener } from '../protectedAudience/utils';
 
 const INFO_ICON_SIZE = 16;
 const INFO_ICON_SPACING = 3;
@@ -80,7 +80,7 @@ const Box = ({
     };
 
     const key = title + description + x + y;
-    utils.addCanvasEventListener('mouseMoved', callback, key);
+    addCanvasEventListener('mouseMoved', callback, key);
   }
 
   p.pop();
