+++ packages/explorable-explanations/src/protectedAudience/app.js
@@ -47,7 +47,6 @@ const app = {
     expandedCircleDiameter: 640,
     minifiedCircleDiameter: 50,
   },
-  utils: {},
   p: null,
   igp: null,
   up: null,
