+++ packages/explorable-explanations/src/protectedAudience/modules/timeline.js
@@ -18,7 +18,7 @@
  */
 import config from '../config.js';
 import app from '../app.js';
-import utils from '../lib/utils.js';
+import * as utils from '../utils';
 import bubbles from './bubbles.js';
 import promiseQueue from '../lib/promiseQueue.js';
 import flow from './flow.js';
