+++ packages/explorable-explanations/src/protectedAudience/utils/canvasEventListener.js
@@ -1,0 +1,28 @@
+/*
+ * Copyright 2024 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+/**
+ * Internal dependencies.
+ */
+import app from '../app';
+
+export const addCanvasEventListener = (
+  event,
+  callback,
+  key,
+  canvasName = 'main'
+) => {
+  app.canvasEventListerners[canvasName][event][key] = callback;
+};
